+++ src/main/java/luckyseven/dart/api/application/auth/AuthenticationService.java
@@ -12,10 +12,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import luckyseven.dart.api.domain.member.entity.Member;
 import luckyseven.dart.api.domain.member.repo.MemberRepository;
-import luckyseven.dart.api.dto.auth.TokenReqDto;
-import luckyseven.dart.api.dto.auth.TokenResDto;
-import luckyseven.dart.api.dto.member.request.LoginReqDto;
-import luckyseven.dart.api.dto.member.response.LoginResDto;
+import luckyseven.dart.dto.auth.TokenReqDto;
+import luckyseven.dart.dto.auth.TokenResDto;
+import luckyseven.dart.dto.member.request.LoginReqDto;
+import luckyseven.dart.dto.member.response.LoginResDto;
 import luckyseven.dart.global.error.exception.BadRequestException;
 import luckyseven.dart.global.error.exception.NotFoundException;
 import luckyseven.dart.global.error.model.ErrorCode;
