+++ src/main/java/luckyseven/dart/api/application/member/MemberService.java
@@ -8,11 +8,10 @@ import lombok.RequiredArgsConstructor;
 import luckyseven.dart.api.domain.auth.AuthUser;
 import luckyseven.dart.api.domain.member.entity.Member;
 import luckyseven.dart.api.domain.member.repo.MemberRepository;
-import luckyseven.dart.api.dto.member.request.MemberUpdateDto;
-import luckyseven.dart.api.dto.member.request.NicknameDuplicationCheckDto;
-import luckyseven.dart.api.dto.member.request.SignUpDto;
-import luckyseven.dart.api.dto.member.response.MemberProfileResDto;
-import luckyseven.dart.global.error.exception.ConflictException;
+import luckyseven.dart.dto.member.request.MemberUpdateDto;
+import luckyseven.dart.dto.member.request.NicknameDuplicationCheckDto;
+import luckyseven.dart.dto.member.request.SignUpDto;
+import luckyseven.dart.dto.member.response.MemberProfileResDto;
 import luckyseven.dart.global.error.exception.NotFoundException;
 import luckyseven.dart.global.error.model.ErrorCode;
