+++ src/main/java/luckyseven/dart/dto/auth/TokenReqDto.java
@@ -1,4 +1,4 @@
-package luckyseven.dart.api.dto.auth;
+package luckyseven.dart.dto.auth;
 
 public record TokenReqDto(
 	String expiredToken
