+++ src/main/java/luckyseven/dart/dto/member/request/LoginReqDto.java
@@ -1,4 +1,4 @@
-package luckyseven.dart.api.dto.member.request;
+package luckyseven.dart.dto.member.request;
 
 import jakarta.validation.constraints.NotBlank;
