+++ src/main/java/luckyseven/dart/dto/member/request/MemberUpdateDto.java
@@ -1,21 +1,12 @@
-package luckyseven.dart.api.dto.member.request;
+package luckyseven.dart.dto.member.request;
 
-import java.time.LocalDate;
-
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
 
 @Builder
-public record SignUpDto(
-	@NotBlank(message = "[❎ ERROR] 이메일을 입력해주세요.")
-	@Email(message = "[❎ ERROR] 올바른 이메일 형식이 아닙니다.")
-	@Size(max = 30, message = "[❎ ERROR] 이메일은 최대 30자까지 가능합니다.")
-	String email,
-
-	@NotBlank(message = "[❎ ERROR] 닉네임을 입력해주세요.")
+public record MemberUpdateDto(
 	@Size(min = 2, max = 10, message = "[❎ ERROR] 닉네임은 2글자에서 10글자 사이여야 합니다.")
 	@Pattern(regexp = "^[A-Za-z\\d가-힣]+$", message = "[❎ ERROR] 닉네임은 한글과 영어만 사용가능합니다.")
 	String nickname,
@@ -28,7 +19,7 @@ public record SignUpDto(
 	)
 	String password,
 
-	LocalDate birthday,
+	String profileImage,
 	String bank,
 	String account,
 	String introduce
